feat(c-api) Move `as_extern` APIs into the `externals` module.
